gibberbot: generate otr_keystore as temp file then move it where it needs to be
This is done to minimize the possibility of leaking the files inadvertently.
For example, if someone selects the SD Card of their phone as the output
device in otrfileconverter-gui, then it should not write the plain text
version of the otr_keystore nor the password encoded in a QRcode .png.