keysync.desktop: remove unneeded declaration of UTF-8 encoding
from lintian:
The Encoding key is now deprecated by the FreeDesktop standard and all
strings are required to be encoded in UTF-8. This desktop entry
explicitly specifies an Encoding of UTF-8, which is harmless but no
longer necessary.

The desktop-file-validate tool in the desktop-file-utils package is
useful for checking the syntax of desktop entries.

More more details, refer to:
http://standards.freedesktop.org/desktop-entry-spec/1.0/apc.html